Backend/requirements.txt
fastapi
uvicorn

pydantic
email-validator
python-dotenv
typing-extensions

passlib[bcrypt]
bcrypt

python-multipart
requests

google-auth
google-auth-oauthlib

aiosmtplib

PyJWT
google-generativeai
jwt